service2/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
tqdm==4.65.0
Pillow==10.3.0
opencv-python-headless==4.10.0.82
pytesseract==0.3.10
requests==2.31.0
transformers==4.11.3
torch==1.10.0
nltk==3.6.3
scikit-learn==0.24.2
unidecode==1.3.4
numpy==1.21.0

albumentations==0.5.2